LayoutTests/imported/w3c/web-platform-tests/content-security-policy/securitypolicyviolation/style-sample-expected.txt


FAIL Inline style blocks should have a sample. assert_equals: expected "p { omg: yay !important; }" but got ""
FAIL Inline style attributes should have a sample. assert_equals: expected "omg: yay !important;" but got ""
